+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   - Added generic `patch` method to the ResourceProvider
   - oneview_resource (generic) support for API300
   - Removed old (unused) helper methods and resource base class
+  - Added support to drive enclosures in API300::Synergy
 
 ### 1.1.0
   - Add support for client ENV variables